+++ services/gateway-service/migrations/008_schools.sql
@@ -1,8 +1,7 @@
 create table if not exists schools (
-  id int(20) primary key auto_increment,
+  id bigint unsigned primary key auto_increment,
   name varchar(255) not null,
   enabled tinyint not null default 1,
   created_at datetime not null default current_timestamp
 );
-
