+++ services/gateway-service/migrations/030_public_config.sql
@@ -1,7 +1,7 @@
 -- Public config tables (per school), with audit
 
 create table if not exists public_config (
-  school_id int primary key,
+  school_id bigint unsigned primary key,
   live tinyint not null default 1,
   rating tinyint not null default 1,
   org_cert tinyint not null default 1,
@@ -14,12 +14,11 @@ create table if not exists public_config (
 );
 
 create table if not exists public_config_audit (
-  id int primary key auto_increment,
-  school_id int not null,
+  id bigint unsigned primary key auto_increment,
+  school_id bigint unsigned not null,
   by_user varchar(128) null,
   at datetime not null default current_timestamp,
   changes_json json null,
   key idx_pca_school_at (school_id, at),
   constraint fk_pca_school foreign key (school_id) references schools(id) on update cascade on delete cascade
 );
-
