+++ services/gateway-service/migrations/051_recreate_supplier_certs_and_unique.sql
@@ -2,8 +2,8 @@
 
 -- 1) Create supplier_certificates if not exists
 CREATE TABLE IF NOT EXISTS supplier_certificates (
-  id int PRIMARY KEY AUTO_INCREMENT,
-  supplier_id int NOT NULL,
+  id bigint unsigned PRIMARY KEY AUTO_INCREMENT,
+  supplier_id bigint unsigned NOT NULL,
   type varchar(64) NOT NULL,
   number varchar(128) NULL,
   authority varchar(255) NULL,
@@ -26,4 +26,3 @@ SET @stmt := (
     'SELECT 1')
 );
 PREPARE s1 FROM @stmt; EXECUTE s1; DEALLOCATE PREPARE s1;
-
