+++ services/gateway-service/migrations/053_create_inbound_attachments.sql
@@ -1,6 +1,6 @@
 -- Attachments for inbound docs (tickets and images)
 CREATE TABLE IF NOT EXISTS inbound_attachments (
-  id int PRIMARY KEY AUTO_INCREMENT,
+  id bigint unsigned PRIMARY KEY AUTO_INCREMENT,
   doc_no varchar(64) NOT NULL,
   type varchar(32) NOT NULL, -- ticket_quarantine, ticket_invoice, ticket_receipt, image
   image_url varchar(255) NOT NULL,
@@ -8,4 +8,3 @@ CREATE TABLE IF NOT EXISTS inbound_attachments (
   KEY idx_inb_att_doc (doc_no),
   KEY idx_inb_att_type (type)
 );
-
