+++ services/gateway-service/migrations/056_create_inv_stocktakes.sql
@@ -1,9 +1,9 @@
 -- Create stocktake records table
 CREATE TABLE IF NOT EXISTS inv_stocktakes (
-  id int primary key auto_increment,
-  school_id int not null,
-  inbound_id int not null,
-  product_id int not null,
+  id bigint unsigned primary key auto_increment,
+  school_id bigint unsigned not null,
+  inbound_id bigint unsigned not null,
+  product_id bigint unsigned not null,
   prev_qty decimal(18,3) not null,
   actual_qty decimal(18,3) not null,
   diff decimal(18,3) not null,
@@ -12,4 +12,3 @@ CREATE TABLE IF NOT EXISTS inv_stocktakes (
   key idx_stocktake_school_at (school_id, at),
   key idx_stocktake_inbound (inbound_id)
 );
-
